new types file for arweave types.
